Refactored test directory structure
