+++ app/src/test/java/org/listenbrainz/android/SocialRepositoryTest.kt
@@ -14,18 +14,18 @@ import org.listenbrainz.android.repository.SocialRepositoryImpl
 import org.listenbrainz.android.service.SocialService
 import org.listenbrainz.android.util.Resource
 import org.listenbrainz.android.util.ResponseError
+import org.listenbrainz.sharedtest.testdata.SocialRepositoryTestData.ErrorUtil.alreadyFollowingError
+import org.listenbrainz.sharedtest.testdata.SocialRepositoryTestData.ErrorUtil.authHeaderNotFoundError
+import org.listenbrainz.sharedtest.testdata.SocialRepositoryTestData.ErrorUtil.cannotFollowSelfError
+import org.listenbrainz.sharedtest.testdata.SocialRepositoryTestData.ErrorUtil.userNotFoundError
+import org.listenbrainz.sharedtest.testdata.SocialRepositoryTestData.testFollowersSuccessData
+import org.listenbrainz.sharedtest.testdata.SocialRepositoryTestData.testFollowingSuccessData
+import org.listenbrainz.sharedtest.testdata.SocialRepositoryTestData.testSearchResult
+import org.listenbrainz.sharedtest.testdata.SocialRepositoryTestData.testSimilarUserSuccessData
 import org.listenbrainz.sharedtest.utils.AssertionUtils
-import org.listenbrainz.sharedtest.utils.EntityTestUtils.ErrorUtil.alreadyFollowingError
-import org.listenbrainz.sharedtest.utils.EntityTestUtils.ErrorUtil.authHeaderNotFoundError
-import org.listenbrainz.sharedtest.utils.EntityTestUtils.ErrorUtil.cannotFollowSelfError
-import org.listenbrainz.sharedtest.utils.EntityTestUtils.ErrorUtil.userNotFoundError
 import org.listenbrainz.sharedtest.utils.EntityTestUtils.testAccessToken
 import org.listenbrainz.sharedtest.utils.EntityTestUtils.testAuthHeader
 import org.listenbrainz.sharedtest.utils.EntityTestUtils.testFamiliarUser
-import org.listenbrainz.sharedtest.utils.EntityTestUtils.testFollowersSuccessData
-import org.listenbrainz.sharedtest.utils.EntityTestUtils.testFollowingSuccessData
-import org.listenbrainz.sharedtest.utils.EntityTestUtils.testSearchResult
-import org.listenbrainz.sharedtest.utils.EntityTestUtils.testSimilarUserSuccessData
 import org.listenbrainz.sharedtest.utils.EntityTestUtils.testSomeOtherUser
 import org.listenbrainz.sharedtest.utils.EntityTestUtils.testUserDNE
 import org.listenbrainz.sharedtest.utils.EntityTestUtils.testUsername
