+++ app/src/test/java/org/listenbrainz/android/YimViewModelTest.kt
@@ -11,14 +11,14 @@ import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
+import org.listenbrainz.sharedtest.mocks.MockAppPreferences
+import org.listenbrainz.sharedtest.mocks.MockYimRepository
 import org.listenbrainz.android.util.Constants.Strings.STATUS_LOGGED_IN
 import org.listenbrainz.android.util.Resource
 import org.listenbrainz.android.viewmodel.YimViewModel
-import org.listenbrainz.sharedtest.mocks.MockAppPreferences
-import org.listenbrainz.sharedtest.mocks.MockYimRepository
 import org.listenbrainz.sharedtest.utils.AssertionUtils.checkYimAssertions
 import org.listenbrainz.sharedtest.utils.EntityTestUtils.testUsername
-import org.listenbrainz.sharedtest.utils.EntityTestUtils.testYimData
+import org.listenbrainz.sharedtest.testdata.YimRepositoryTestData.testYimData
 
 class YimViewModelTest{
